Edit a map's global variables in its .gam

+++ src/cli/Endings.cpp
@@ -24,7 +24,7 @@ namespace {
         if (gam == nullptr || gvar < 0 || static_cast<std::size_t>(gvar) >= gam->gvarCount()) {
             return {};
         }
-        return gam->gvarKey(gvar);
+        return gam->gvarKey(static_cast<std::size_t>(gvar));
     }
 
 } // namespace
